requirements.txt
streamlit
langchain_community
langchain_groq
numpy
scikit-learn
nltk
requests
python-dotenv
